spatial_narratives_api/requirements.txt
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
json5==0.9.14
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
psycopg2-binary==2.9.9